Simplify; no destructuring
Don't bother destructuring the tuples in the rope.  Just use .0 and .1 to access the elements.

Also, we don't need a local copy of the current "tail" inside the loop in step_head.  Use a mutuable reference to update it directly.  We *do* need a copy of the current "head" since we can't have an immutable reference to one element and a mutable reference to another at the same time.