Day 24, Part 1 is MUCH faster
Don't store the positions of the blizzards in the State.  Therefore,
we don't have to calculate new positions for every blizzard in every
state, and don't have to collect into a Vec and HashSet.  Instead,
we can directly compute the position of a blizzard at a given time
with some simple math.

There is no need to compute the positions of all blizzards to detect
a collision.  Since a blizzard stays in a single row or column, we
only need to check blizzards in the row and column of interest.  That
reduces the total number of blizzards to check.