Day 18, part 2: switch to pathfinding's dfs
Use the dfs() function from the pathfinding crate, rather than my own hand-rolled implementation.  Since it uses closures, it was easy to maintain a set of visited points for updating the cache.

The only gotcha was that I needed to prevent generating the same neighbors multiple times (from multiple invocations of the closure).  This was critical to keeping performace in line with my custom dfs.